+++ src/Rules/VatNumberExist.php
@@ -16,7 +16,7 @@ class VatNumberExist implements ValidationRule
     public function validate(string $attribute, mixed $value, Closure $fail): void
     {
         if (! $this->isValid($value)) {
-            $fail(__('VAT number :attribute  not exist.'));
+            $fail(__('laravelVatEuValidator::validation.vat_number_exist', ['attribute' => $attribute]));
         }
     }
 }